The squirrely case of Jacques
Every now and then, usually between eight and ten in the evening,
Jacques finds himself transforming into a small furry rodent with a
bushy tail. For a while, he believed that it happened only on days when
he had touched trees. So he stopped touching trees entirely and even
avoided going near them. But the problem persisted.Switching to a more
scientific approach, Jacques intends to start keeping a daily log of
everything he did that day and whether he changed form. With this data
he hopes to narrow down the conditions that trigger the transformations.
This program represents the computation involved in finding out the
reason behind his transformation.